load_balancer_service/requirements.txt
pika>=1.3,<2.0
python-dotenv>=1.0,<2.0
